move blinding.py to own file and add secure_secret_path method
