+++ buidl/__init__.py
@@ -1,6 +1,7 @@
 # TODO: specify what specifically to import
 from .bcur import *
 from .bech32 import *
+from .blinding import *
 from .block import *
 from .bloomfilter import *
 from .ecc import *
