+++ buidl/hd.py
@@ -686,50 +686,6 @@ def ltrim_path(bip32_path, depth):
     return "m/" + "/".join(to_return)
 
 
-def combine_bip32_paths(first_path, second_path):
-    for bip32_path in (first_path, second_path):
-        if not is_valid_bip32_path(bip32_path):
-            raise ValueError(f"Invalid bip32 path: {bip32_path}")
-
-    # be forgiving
-    first_path = first_path.lower().strip().replace("'", "h").replace("//", "/")
-    second_path = second_path.lower().strip().replace("'", "h").replace("//", "/")
-
-    if first_path == "m":
-        return second_path
-
-    if second_path == "m":
-        return first_path
-
-    # Trim of leading "m/" from second path:
-    return f"{first_path}/{second_path[2:]}"
-
-
-def blind_xpub(starting_xpub, starting_path, secret_path):
-    """
-    Blind a starting_xpub with a given (and unverifiable) path, using a secret path.
-
-    Return the complete (combined) bip32 path, and
-    """
-
-    starting_xpub_obj = HDPublicKey.parse(starting_xpub)
-    # Note that we cannot verify the starting path, so it is essential that at least this safety check is accurate
-    if starting_xpub_obj.depth != starting_path.count("/"):
-        raise ValueError(
-            f"starting_xpub_obj.depth {starting_xpub_obj.depth} != starting_path depth {starting_path.count('/')}"
-        )
-
-    # This will automatically use the version byte that was parsed in the previous step
-    blinded_child_xpub = starting_xpub_obj.traverse(secret_path).xpub()
-    blinded_full_path = combine_bip32_paths(
-        first_path=starting_path, second_path=secret_path
-    )
-    return {
-        "blinded_child_xpub": blinded_child_xpub,
-        "blinded_full_path": blinded_full_path,
-    }
-
-
 def parse_key_record(key_record_str):
     """
     A key record will look something like this:
